Remove casting of user and item IDs in Spark evaluation

+++ recommenders/evaluation/spark_evaluation.py
@@ -99,13 +99,13 @@ class SparkRatingEvaluation:
             raise ValueError("Schema of rating_pred not valid. Missing Prediction Col")
 
         self.rating_true = self.rating_true.select(
-            col(self.col_user).cast("double"),
-            col(self.col_item).cast("double"),
+            col(self.col_user),
+            col(self.col_item),
             col(self.col_rating).cast("double").alias("label"),
         )
         self.rating_pred = self.rating_pred.select(
-            col(self.col_user).cast("double"),
-            col(self.col_item).cast("double"),
+            col(self.col_user),
+            col(self.col_item),
             col(self.col_prediction).cast("double").alias("prediction"),
         )
